refactor(validator): 패키지간 양방향 의존성 관계 끊기 위해 의존성 역전 원리 사용
